chore: refactor hooks files structure
