+++ src/hooks/api/manga/useNewPopularTitles.ts
@@ -1,5 +1,5 @@
-import { useQuery } from '@tanstack/react-query';
 import { getManga, MangaExpansions } from '@api/manga-api';
+import { useQuery } from '@tanstack/react-query';
 import { ContentRating } from '@types';
 import { convertDate, manipulateDate } from '@utils';
 
@@ -31,25 +31,4 @@ const useNewPopularTitles = (enabled?: boolean) => {
   return query;
 };
 
-const useRecentlyAdded = (enabled?: boolean) => {
-  const query = useQuery({
-    enabled,
-    queryKey: ['recently-added'],
-    queryFn: () =>
-      getManga({
-        limit: 15,
-        hasAvailableChapters: true,
-        includes: [MangaExpansions.COVER],
-        order: { createdAt: 'desc' },
-        contentRating: [
-          ContentRating.EROTICA,
-          ContentRating.SUGGESTIVE,
-          ContentRating.SAFE,
-        ],
-      }),
-  });
-
-  return query;
-};
-
-export { useNewPopularTitles, useRecentlyAdded };
+export { useNewPopularTitles };
