+++ src/hooks/index.ts
@@ -1,4 +1,5 @@
 export { useHydration } from './useHydration';
-export { useNewPopularTitles, useRecentlyAdded } from './api/manga-hooks';
-export { useSeasonalList } from './api/custom-list-hooks';
-export { useLatestUploads } from './api/chapter-hooks';
+export { useNewPopularTitles } from './api/manga/useNewPopularTitles';
+export { useRecentlyAddedTitles } from './api/manga/useRecentlyAddedTitles';
+export { useSeasonalList } from './api/custom-list/useSeasonalList';
+export { useLatestUploads } from './api/chapter/useLatestUploads';
